requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Web Framework
streamlit>=1.28.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0
